chore: expanded the version scope for the TextMesh Pro package
Currently, the QvPen project specified the TextMesh Pro version ^2.1.4 is
specified as the required  version, but Unity 2022 requires migration to v3.
Since the latest version of the current v3 series is 3.0.6, the intent is
to include this version in the allowed range so that it will also work with
Unity 2022.

see: https://docs.unity3d.com/ja/2022.3/Manual/com.unity.textmeshpro.html